requirements.txt
pygame
requests
psutil
openai
python-dotenv
SpeechRecognition
pyaudio
